requirements.txt
streamlit
pandas
gspread
oauth2client
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai